pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.18"
description = ""
authors = ["Artem Ivanov <wertart@gmail.com>"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^6.2.4"
pytest-cov = "^2.12.1"
PyYAML = "^5.4.1"
pip = "^21.1.3"
install = "^1.3.4"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.gendiff:main"
